Fix engine timer alignment

+++ game/library/structure/engine_timer.h
@@ -6,7 +6,7 @@
 struct engine_timer
 {
     readonly uint32 frame;
-    readonly uint16 timer_handle;
+    readonly int16 timer_handle;
     padding(2)
     readonly uint32 timer_count_current;
     readonly uint32 timer_count_delta;
@@ -15,8 +15,8 @@ struct engine_timer
     readonly uint32 pause_time;
     readonly float32 frame_length;
     padding(4)
-    readonly struct { readonly uint64_t low, high; } time_real;
-    readonly struct { readonly uint64_t low, high; } time_pause;
+    readonly struct { readonly uint32_t low, high; } time_real;
+    readonly struct { readonly uint32_t low, high; } time_pause;
     readonly uint32 ticks_per_ms;
     padding(4)
 };
